requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary==2.9.9
supabase

